+++ lib/viewmodel.dart
@@ -2,3 +2,4 @@ library viewmodel;
 
 export 'package:viewmodel/src/viewmodel.dart';
 export 'package:viewmodel/src/viewmodel_mixins.dart';
+export 'package:viewmodel/src/viewmodel_exceptions.dart';